+++ api/src/services/createTransport.js
@@ -1,0 +1,17 @@
+const nodemailer = require('nodemailer');
+
+const createNodemailerTransport = async () => {
+  return nodemailer.createTransport({
+    host: 'smtp.gmail.com',
+    port: 587,
+    secure: false,
+    auth: {
+      user: 'verdevolver@gmail.com',
+      pass: 'hngpoqxjettmhpiv',
+    },
+  });
+};
+
+module.exports = {
+  createNodemailerTransport,
+};
